fix bug in implementation of Array.prototype.flatMap
Previously, if you were to call the flatMap function with a number higher than 1 as the second argument, the function would recursively call the passed lambda on the array however many times you specified in the second argument. This could easily cause a bug since most mapper functions are only supposed to be ran once on an array. For example, if you called `[[[1,2,3]]].flatMap(x => x * 2, 3)`, you would get back `[1, 16, 18]` instead of [1, 4, 6]`. This is because the array would get doubled 3 times instead of just once. What I did to fix this bug was remove the second parameter (levels) from the function signature and remove the recursion. As a result, you can only flatten an array one level deep.